refactoring day04
