+++ 2019/resources/day04/day04.txt
@@ -1,0 +1,1 @@
+178416-676461
