+++ 2019/src/lib.rs
@@ -1,4 +1,6 @@
+pub mod utils;
+
 pub mod day01;
 pub mod day02;
 pub mod day03;
-pub mod utils;
+pub mod day04;
