+++ 2019/tests/day04_test.rs
@@ -1,0 +1,14 @@
+#[cfg(test)]
+mod tests {
+    use aoc2019::day04;
+
+    #[test]
+    fn part1_works() {
+        assert_eq!(1650, day04::part1());
+    }
+
+    #[test]
+    fn part2_works() {
+        assert_eq!(1129, day04::part2());
+    }
+}
